Remove `App` type
