+++ godoc-current.txt
@@ -317,161 +317,6 @@ type AfterFunc func(*Context) error
     AfterFunc is an action that executes after any subcommands are run and have
     finished. The AfterFunc is run even if Action() panics.
 
-type App struct {
-	// The name of the program. Defaults to path.Base(os.Args[0])
-	Name string
-	// Full name of command for help, defaults to Name
-	HelpName string
-	// Description of the program.
-	Usage string
-	// Text to override the USAGE section of help
-	UsageText string
-	// Description of the program argument format.
-	ArgsUsage string
-	// Version of the program
-	Version string
-	// Description of the program
-	Description string
-	// DefaultCommand is the (optional) name of a command
-	// to run if no command names are passed as CLI arguments.
-	DefaultCommand string
-	// List of commands to execute
-	Commands []*Command
-	// List of flags to parse
-	Flags []Flag
-	// Boolean to enable shell completion commands
-	EnableShellCompletion bool
-	// Shell Completion generation command name
-	ShellCompletionCommandName string
-	// Boolean to hide built-in help command and help flag
-	HideHelp bool
-	// Boolean to hide built-in help command but keep help flag.
-	// Ignored if HideHelp is true.
-	HideHelpCommand bool
-	// Boolean to hide built-in version flag and the VERSION section of help
-	HideVersion bool
-
-	// An action to execute when the shell completion flag is set
-	ShellComplete ShellCompleteFunc
-	// An action to execute before any subcommands are run, but after the context is ready
-	// If a non-nil error is returned, no subcommands are run
-	Before BeforeFunc
-	// An action to execute after any subcommands are run, but after the subcommand has finished
-	// It is run even if Action() panics
-	After AfterFunc
-	// The action to execute when no subcommands are specified
-	Action ActionFunc
-	// Execute this function if the proper command cannot be found
-	CommandNotFound CommandNotFoundFunc
-	// Execute this function if a usage error occurs
-	OnUsageError OnUsageErrorFunc
-	// Execute this function when an invalid flag is accessed from the context
-	InvalidFlagAccessHandler InvalidFlagAccessFunc
-	// List of all authors who contributed (string or fmt.Stringer)
-	Authors []any // TODO: ~string | fmt.Stringer when interface unions are available
-	// Copyright of the binary if any
-	Copyright string
-	// Reader reader to write input to (useful for tests)
-	Reader io.Reader
-	// Writer writer to write output to
-	Writer io.Writer
-	// ErrWriter writes error output
-	ErrWriter io.Writer
-	// ExitErrHandler processes any error encountered while running an App before
-	// it is returned to the caller. If no function is provided, HandleExitCoder
-	// is used as the default behavior.
-	ExitErrHandler ExitErrHandlerFunc
-	// Other custom info
-	Metadata map[string]interface{}
-	// Carries a function which returns app specific info.
-	ExtraInfo func() map[string]string
-	// CustomAppHelpTemplate the text template for app help topic.
-	// cli.go uses text/template to render templates. You can
-	// render custom help text by setting this variable.
-	CustomAppHelpTemplate string
-	// SliceFlagSeparator is used to customize the separator for SliceFlag, the default is ","
-	SliceFlagSeparator string
-	// DisableSliceFlagSeparator is used to disable SliceFlagSeparator, the default is false
-	DisableSliceFlagSeparator bool
-	// Boolean to enable short-option handling so user can combine several
-	// single-character bool arguments into one
-	// i.e. foobar -o -v -> foobar -ov
-	UseShortOptionHandling bool
-	// Enable suggestions for commands and flags
-	Suggest bool
-	// Allows global flags set by libraries which use flag.XXXVar(...) directly
-	// to be parsed through this library
-	AllowExtFlags bool
-	// Treat all flags as normal arguments if true
-	SkipFlagParsing bool
-	// Flag exclusion group
-	MutuallyExclusiveFlags []MutuallyExclusiveFlags
-	// Use longest prefix match for commands
-	PrefixMatchCommands bool
-	// Custom suggest command for matching
-	SuggestCommandFunc SuggestCommandFunc
-
-	// Has unexported fields.
-}
-    App is the main structure of a cli application.
-
-func (a *App) Command(name string) *Command
-    Command returns the named command on App. Returns nil if the command does
-    not exist
-
-func (a *App) Run(arguments []string) error
-    Run is the entry point to the cli app. Parses the arguments slice and routes
-    to the proper flag/args combination
-
-func (a *App) RunContext(ctx context.Context, arguments []string) (err error)
-    RunContext is like Run except it takes a Context that will be passed to
-    its commands and sub-commands. Through this, you can propagate timeouts and
-    cancellation requests
-
-func (a *App) Setup()
-    Setup runs initialization code to ensure all data structures are ready
-    for `Run` or inspection prior to `Run`. It is internally called by `Run`,
-    but will return early if setup has already happened.
-
-func (a *App) ToFishCompletion() (string, error)
-    ToFishCompletion creates a fish completion string for the `*App` The
-    function errors if either parsing or writing of the string fails.
-
-func (a *App) ToMan() (string, error)
-    ToMan creates a man page string for the `*App` The function errors if either
-    parsing or writing of the string fails.
-
-func (a *App) ToManWithSection(sectionNumber int) (string, error)
-    ToMan creates a man page string with section number for the `*App` The
-    function errors if either parsing or writing of the string fails.
-
-func (a *App) ToMarkdown() (string, error)
-    ToMarkdown creates a markdown string for the `*App` The function errors if
-    either parsing or writing of the string fails.
-
-func (a *App) ToTabularMarkdown(appPath string) (string, error)
-    ToTabularMarkdown creates a tabular markdown documentation for the `*App`.
-    The function errors if either parsing or writing of the string fails.
-
-func (a *App) ToTabularToFileBetweenTags(appPath, filePath string, startEndTags ...string) error
-    ToTabularToFileBetweenTags creates a tabular markdown documentation for
-    the `*App` and updates the file between the tags in the file. The function
-    errors if either parsing or writing of the string fails.
-
-func (a *App) VisibleCategories() []CommandCategory
-    VisibleCategories returns a slice of categories and commands that are
-    Hidden=false
-
-func (a *App) VisibleCommands() []*Command
-    VisibleCommands returns a slice of the Commands with Hidden=false
-
-func (a *App) VisibleFlagCategories() []VisibleFlagCategory
-    VisibleFlagCategories returns a slice containing all the categories with the
-    flags they contain
-
-func (a *App) VisibleFlags() []Flag
-    VisibleFlags returns a slice of the Flags with Hidden=false
-
 type Args interface {
 	// Get returns the nth argument, or else a blank string
 	Get(n int) string
