+++ internal/example-cli/example-cli.go
@@ -3,9 +3,11 @@
 package main
 
 import (
+	"context"
+
 	"github.com/urfave/cli/v3"
 )
 
 func main() {
-	_ = (&cli.App{}).Run([]string{""})
+	_ = (&cli.Command{}).Run(context.Background(), []string{""})
 }
